requirements.txt
# وابستگی‌های اصلی
torch
torchaudio
transformers
openai-whisper
streamlit
pyaudio
numpy
scipy
librosa
soundfile

# وابستگی‌های اضافی برای فاز اول
# wave, tempfile, threading, io, os, time ماژول‌های built-in هستند

# وابستگی‌های GPU (اختیاری)
# CUDA toolkit باید جداگانه نصب شود
# nvidia-ml-py3

# وابستگی‌های توسعه
pytest
black
flake8
